Refactor structure of chat page
