+++ src/pages/chat/components/ChatDialog/components/ChatDialogInputRow/components/index.js
@@ -1,0 +1,1 @@
+export { default as ChatImageUpload } from './ChatImageUpload';
