+++ src/pages/chat/components/ChatDialog/components/ChatDialogMessages/components/ChatMessageSection/components/ChatBubbleForImage/index.js
@@ -1,5 +1,7 @@
 import React from 'react';
 import styled from 'styled-components';
+
+// constants
 import { colors } from '@constants/colors';
 
 // note that the max height of the image is set to 80% of the chat dialog messages' height,
