+++ src/pages/chat/components/ChatDialog/components/ChatDialogMessages/components/ChatMessageSection/components/index.js
@@ -1,0 +1,3 @@
+export { default as ChatBubbleForCalendar } from './ChatBubbleForCalendar';
+export { default as ChatBubbleForImage } from './ChatBubbleForImage';
+export { default as ChatBubbleForText } from './ChatBubbleForText';
