+++ src/pages/chat/components/ChatDialog/components/ChatDialogMessages/components/index.js
@@ -1,0 +1,2 @@
+export { LeftMessageSection, RightMessageSection } from './ChatMessageSection';
+export { default as NewChatTipsForWish } from './NewChatTipsForWish';
