+++ src/pages/chat/components/ChatDialog/components/index.js
@@ -1,0 +1,4 @@
+export { default as ChatDialogInputRow } from './ChatDialogInputRow';
+export { default as ChatDialogMessages } from './ChatDialogMessages';
+export { default as ChatDialogUserRow } from './ChatDialogUserRow';
+export { default as ChatDialogViewPostRow } from './ChatDialogViewPostRow';
