+++ src/pages/chat/components/ListOfChats/index.js
@@ -1,16 +1,31 @@
 import React, { useState, useEffect, useContext } from 'react';
-import { Stack, TileGroup, Loading } from '@kiwicom/orbit-components/lib';
-import ChatWithUserCard from '../../card/ChatWithUserCard';
-import api from '@api';
 import styled, { css } from 'styled-components';
 import media from '@kiwicom/orbit-components/lib/utils/mediaQuery';
+
+// components
+import { Stack, TileGroup, Loading } from '@kiwicom/orbit-components/lib';
+import ChatWithUserCard from '@components/card/ChatWithUserCard';
+import InfiniteScroll from '@components/scroller/InfiniteScroller';
+
+// constants and utils
+import api from '@api';
 import { colors } from '@constants/colors';
 import { MODIFIED, ADDED } from '@constants/chatSubscriptionChange';
-import InfiniteScroll from '../../scroller/InfiniteScroller';
 import { USER_CHATS_BATCH_SIZE } from '@api/constants';
-import ChatContext from '../context';
-import { setSelectedChatId, setIsNewChat } from '../actions';
-import { getSelectedChatId, getIsNewChat, getUser, getIsViewingChatsForMyPost, getPostId } from '../selectors';
+
+// context
+import {
+  ChatContext,
+  // selectors
+  getSelectedChatId,
+  getIsNewChat,
+  getUser,
+  getIsViewingChatsForMyPost,
+  getPostId,
+  // actions
+  setSelectedChatId,
+  setIsNewChat,
+} from '../../context';
 
 const ListOfChatsContainer = styled.div`
   min-width: 200px;
