+++ src/pages/chat/components/index.js
@@ -1,0 +1,2 @@
+export { default as ChatDialog } from './ChatDialog';
+export { default as ListOfChats } from './ListOfChats';
